Update api.js
I've been having a problem with Better Looting and Harvesting (harvester mod) and so this is a fix to the incoming string not being in the expected format for ItemPiles.

I'm guessing there's a better solution to this than what I did.

I appreciate you making the v12 update btw. Saved me a lot of headache trying to figure it out myself.